Fix and merge idealabs
