+++ cypress/e2e/ideyalabs/organizations/delete-organization.cy.js
@@ -1,0 +1,43 @@
+import devTeams from '../../../support/dictionary/devTeams';
+import testTypes from '../../../support/dictionary/testTypes';
+import organizations from '../../../support/fragments/organizations/organizations';
+import getRandomStringCode from '../../../support/utils/genereteTextCode';
+import getRandomPostfix from '../../../support/utils/stringTools';
+
+const organizationName = `AutoTest_${getRandomPostfix()}`;
+const organizationCode = getRandomStringCode(6);
+
+const newOrganization = {
+  status: 'Active',
+  name: organizationName,
+  code: organizationCode,
+};
+
+const searchByCode = {
+  dropdown: 'Code',
+  code: organizationCode,
+};
+
+describe('ui-organizations: Organizations creation', () => {
+  before('Login', () => {
+    cy.login(Cypress.env('diku_login'), Cypress.env('diku_password'));
+  });
+
+  after('Verifying deleted test data', () => {
+    organizations.searchByParameters(searchByCode.dropdown, searchByCode.code);
+    organizations.checkZeroSearchResultsHeader();
+  });
+
+  it(
+    'C674 Delete existing organization record (thunderjet)',
+    { tags: [testTypes.extendedPath, devTeams.thunderjet] },
+    () => {
+      organizations.searchByParameters(
+        searchByCode.dropdown,
+        searchByCode.code
+      );
+      organizations.selectOrganization(newOrganization.name);
+      organizations.deleteOrganization();
+    }
+  );
+});
